+++ libs/sdmx-toolkit/src/models/index.ts
@@ -3,9 +3,9 @@ export * from './dimension-group';
 export * from './series-filter';
 export * from './splitted-urn';
 export * from './data';
-export * from './structural-metadata';
-export * from './structural-metadata-base';
-export * from './structural-metadata-v3';
+export * from './structural-metadata/structural-metadata';
+export * from './structural-metadata/structural-metadata-base';
+export * from './structural-metadata/structural-metadata-v3';
 export * from './structural-metadata/codelist';
 export * from './structural-metadata/common-artefact-properties';
 export * from './structural-metadata/concept-scheme';
