+++ libs/sdmx-toolkit/src/models/structural-metadata/codelist.ts
@@ -1,5 +1,5 @@
 import { CommonArtefactProperty } from './common-artefact-properties';
-import { ElementBase } from '../structural-metadata-base';
+import { ElementBase } from './structural-metadata-base';
 
 export interface Codelist extends CommonArtefactProperty {
   codes?: Code[];
