+++ libs/sdmx-toolkit/src/models/structural-metadata/common-artefact-properties.ts
@@ -1,4 +1,4 @@
-import { Annotation, ElementBase, Link } from '../structural-metadata-base';
+import { Annotation, ElementBase, Link } from './structural-metadata-base';
 
 export interface CommonArtefactProperty extends ElementBase {
   agencyID?: string;
