+++ libs/sdmx-toolkit/src/models/structural-metadata/concept-scheme.ts
@@ -1,6 +1,6 @@
 import { CommonArtefactProperty } from './common-artefact-properties';
 import { Representation } from './representation';
-import { ElementBase } from '../structural-metadata-base';
+import { ElementBase } from './structural-metadata-base';
 
 export interface ConceptScheme extends CommonArtefactProperty {
   concepts?: Concept[];
