+++ libs/sdmx-toolkit/src/models/structural-metadata/structural-metadata-base.ts
